Big reorganization of all the player files
