+++ lib/models/club/page/club_page.dart
@@ -11,7 +11,7 @@ import 'package:opengoalz/models/profile.dart';
 import 'package:opengoalz/models/league/league.dart';
 import 'package:opengoalz/models/player/class/player.dart';
 import 'package:opengoalz/constants.dart';
-import 'package:opengoalz/models/player/players_page.dart';
+import 'package:opengoalz/models/player/pages/players_page.dart';
 import 'package:opengoalz/pages/user_page/user_page.dart';
 import 'package:opengoalz/postgresql_requests.dart';
 import 'package:opengoalz/provider_theme_app.dart';
