+++ lib/models/events/event.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/players_page.dart';
+import 'package:opengoalz/models/player/pages/players_page.dart';
 import 'package:opengoalz/models/playerSearchCriterias.dart';
 
 class GameEvent {
