+++ lib/models/player/class/player.dart
@@ -6,10 +6,10 @@ import 'package:opengoalz/functions/AgeAndBirth.dart';
 import 'package:opengoalz/models/club/class/club.dart';
 import 'package:opengoalz/models/game/gamePlayerStatsAll.dart';
 import 'package:opengoalz/models/multiverse/multiverse.dart';
-import 'package:opengoalz/models/player/playerTransferListTile.dart';
+import 'package:opengoalz/models/player/transfer/playerTransferListTile.dart';
 import 'package:opengoalz/models/player/playerExpensesListTile.dart';
-import 'package:opengoalz/models/player/playerWidgets.dart';
-import 'package:opengoalz/models/player/player_embodied_listtile.dart';
+import 'package:opengoalz/models/player/widgets/playerWidgets.dart';
+import 'package:opengoalz/models/player/embodied/player_embodied_listtile.dart';
 import 'package:opengoalz/models/playerPoaching/player_poaching.dart';
 import 'package:opengoalz/models/playerSearchCriterias.dart';
 import 'package:opengoalz/models/playerFavorite/player_favorite.dart';
@@ -17,11 +17,11 @@ import 'package:opengoalz/models/playerStatsBest.dart';
 import 'package:opengoalz/models/profile.dart';
 import 'package:opengoalz/models/transfer_bid.dart';
 import 'package:opengoalz/constants.dart';
-import 'package:opengoalz/models/player/players_page.dart';
+import 'package:opengoalz/models/player/pages/players_page.dart';
 import 'package:opengoalz/widgets/countryListTile.dart';
 import 'package:opengoalz/widgets/graphWidget.dart';
 import 'package:collection/collection.dart';
-import 'package:opengoalz/models/player/transfers_embodied_players_offer.dart';
+import 'package:opengoalz/models/player/embodied/transfers_embodied_players_offer.dart';
 
 part 'player_widget_helper.dart';
