+++ lib/models/player/dialogs/playerTrainingCoefDialogBox.dart
