+++ lib/models/player/dialogs/playerTransferBidDialogBox.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:opengoalz/functions/loadingCircularAndText.dart';
 import 'package:opengoalz/models/club/others/clubCashListTile.dart';
-import 'package:opengoalz/models/player/playerTransferListTile.dart';
+import 'package:opengoalz/models/player/transfer/playerTransferListTile.dart';
 import 'package:opengoalz/models/transfer_bid.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:opengoalz/constants.dart';
