+++ lib/models/player/dialogs/player_training_dialog.dart
