+++ lib/models/player/embodied/player_embodied_listtile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:opengoalz/constants.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/player_embodied_offers_button.dart';
+import 'package:opengoalz/models/player/widgets/player_embodied_offers_button.dart';
 import 'package:opengoalz/models/profile.dart';
 
 class PlayerCardEmbodiedListTile extends StatelessWidget {
