+++ lib/models/player/transfer/player_transfer_list_tile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:opengoalz/constants.dart';
 import 'package:opengoalz/models/club/others/getClubNameWidget.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/playerTransferBidDialogBox.dart';
+import 'package:opengoalz/models/player/dialogs/playerTransferBidDialogBox.dart';
 import 'package:opengoalz/widgets/tickingTime.dart';
 
 class PlayerCardTransferWidget extends StatelessWidget {
