+++ lib/models/player/widgets/list_tiles/player_expenses_list_tile.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:opengoalz/constants.dart';
 import 'package:opengoalz/extensionBuildContext.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/playerWidgets.dart';
+import 'package:opengoalz/models/player/widgets/get_player_history_graph.dart';
+import 'package:opengoalz/models/player/widgets/player_widgets.dart';
 import 'package:opengoalz/postgresql_requests.dart';
 import 'package:opengoalz/provider_user.dart';
 import 'package:provider/provider.dart';
