+++ lib/models/player/widgets/player_widgets.dart
@@ -2,13 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:opengoalz/constants.dart';
 import 'package:opengoalz/functions/AgeAndBirth.dart';
-import 'package:opengoalz/functions/loadingCircularAndText.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/playerNotesDialogBox.dart';
-import 'package:opengoalz/models/player/playerShirtNumberDialogBox.dart';
+import 'package:opengoalz/models/player/dialogs/player_notes_dialog_box.dart';
+import 'package:opengoalz/models/player/dialogs/player_shirt_number_dialog_box.dart';
+import 'package:opengoalz/models/player/pages/player_history_page.dart';
 import 'package:opengoalz/provider_user.dart';
-import 'package:opengoalz/widgets/graphWidget.dart';
-import 'package:opengoalz/widgets/playerHistoryPage.dart';
 import 'package:provider/provider.dart';
 
 Widget playerShirtNumberIcon(BuildContext context, Player player) {
@@ -80,84 +78,6 @@ Widget playerSmallNotesIcon(BuildContext context, Player player) {
         ],
       ),
     ),
-  );
-}
-
-Widget getPlayerHistoryGraph(
-    BuildContext context, int id, List<String> fieldsToPlot, String title) {
-  return FutureBuilder<List<Map>>(
-    future: supabase
-        .from('players_history_stats')
-        .select('created_at, ${fieldsToPlot.join(", ")}')
-        .eq('id_player', id)
-        .order('created_at', ascending: true)
-        .then((response) => response),
-    builder: (context, snapshot) {
-      if (snapshot.connectionState == ConnectionState.waiting) {
-        return loadingCircularAndText('Loading player history...');
-      } else if (snapshot.hasError) {
-        return AlertDialog(
-          title: Text('Error'),
-          content: Text(
-            'Failed to load player history: ${snapshot.error}',
-            style: TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: Text('OK'),
-            ),
-          ],
-        );
-      } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
-        return AlertDialog(
-          title: Text('Error'),
-          content: Text(
-            'No data available for player history.',
-            style: TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.of(context).pop(),
-              child: Text('OK'),
-            ),
-          ],
-        );
-      }
-
-      final historyData = snapshot.data!;
-
-      // Check for null values in the required fields
-      for (var field in fieldsToPlot) {
-        if (historyData.any((item) => item[field] == null)) {
-          return AlertDialog(
-            title: Text('Data Error'),
-            content: Text(
-              'Error: Missing data for field "$field".',
-              style: TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
-            ),
-            actions: [
-              TextButton(
-                onPressed: () => Navigator.of(context).pop(),
-                child: Text('OK'),
-              ),
-            ],
-          );
-        }
-      }
-
-      final chartData = ChartData(
-        title: title,
-        yValues: fieldsToPlot.map((field) {
-          return historyData
-              .map((item) => (item[field] as num).toDouble())
-              .toList();
-        }).toList(),
-        typeXAxis: XAxisType.weekHistory,
-      );
-
-      return ChartDialogBox(chartData: chartData);
-    },
   );
 }
