+++ lib/pages/transfer_page.dart
@@ -3,7 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:opengoalz/functions/loadingCircularAndText.dart';
 import 'package:opengoalz/models/club/others/clubCashListTile.dart';
-import 'package:opengoalz/models/player/playerCard_Main.dart';
+import 'package:opengoalz/models/player/dialogs/playerSearchDialogBox.dart';
+import 'package:opengoalz/models/player/widgets/playerCard_Main.dart';
 import 'package:opengoalz/models/profile.dart';
 import 'package:opengoalz/widgets/goBackToolTip.dart';
 import 'package:opengoalz/widgets/tab_widget_with_icon.dart';
@@ -13,10 +14,9 @@ import 'package:opengoalz/models/playerSearchCriterias.dart';
 import 'package:opengoalz/models/transfer_bid.dart';
 import 'package:opengoalz/provider_user.dart';
 import 'package:opengoalz/models/player/class/player.dart';
-import 'package:opengoalz/models/player/players_page.dart';
+import 'package:opengoalz/models/player/pages/players_page.dart';
 import 'package:opengoalz/widgets/appDrawer.dart';
 import 'package:opengoalz/widgets/max_width_widget.dart';
-import 'package:opengoalz/models/player/playerSearchDialogBox.dart';
 import 'package:provider/provider.dart';
 
 import '../constants.dart';
