+++ lib/widgets/staff_detail_tab.dart
@@ -4,10 +4,10 @@ import 'package:opengoalz/models/club/class/club.dart';
 import 'package:opengoalz/models/club/others/getClubNameWidget.dart';
 import 'package:opengoalz/models/player/class/player.dart';
 import 'package:opengoalz/constants.dart';
-import 'package:opengoalz/models/player/playerStaffCoefListTile.dart';
 import 'package:opengoalz/models/player/playerExpensesListTile.dart';
-import 'package:opengoalz/models/player/playerWidgets.dart';
-import 'package:opengoalz/models/player/players_page.dart';
+import 'package:opengoalz/models/player/pages/players_page.dart';
+import 'package:opengoalz/models/player/widgets/playerStaffCoefListTile.dart';
+import 'package:opengoalz/models/player/widgets/playerWidgets.dart';
 import 'package:opengoalz/models/playerSearchCriterias.dart';
 import 'package:opengoalz/postgresql_requests.dart';
 import 'package:opengoalz/widgets/countryListTile.dart';
